Allow fractional mindiff/startdiff and configurable donation addresses
mindiff and startdiff now accept floating-point values in ckpool.conf.
Values less than 1 are passed through as-is, enabling sub-1 difficulty
for low-hashrate or testing environments. Values greater than 1 are
floored to the nearest integer, preserving existing behaviour for all
current configurations.

Add json_get_num_as_double() helper that accepts both JSON integer and
real nodes, so existing configs with integer-typed mindiff/startdiff
continue to work without modification.

The donation recipient addresses (donaddress, tndonaddress, rtdonaddress)
are now configurable via ckpool.conf. The hardcoded developer addresses
remain as fallback defaults when the keys are absent.

Changes:
- src/ckpool.h: mindiff/startdiff int64_t -> double; declare
  json_get_num_as_double()
- src/ckpool.c: add json_get_num_as_double(); parse mindiff/startdiff
  with new helper; apply floor() for values > 1; parse donaddress,
  tndonaddress, rtdonaddress from config with hardcoded fallbacks
- src/stratifier.c: update all mindiff/startdiff usages for double
  type; guard suggest_diff path with >= 1.0 check; cast to int64_t
  where client->diff or JSON int64 is required
- src/generator.c: cast ckp->mindiff to int64_t in JSON_CPACK
- ckpool.conf, README, README-SOLOMINING: document new config keys

+++ src/generator.c
@@ -2037,7 +2037,7 @@ static void suggest_diff(ckpool_t *ckp, connsock_t *cs, proxy_instance_t *proxy)
 	JSON_CPACK(req, "{s:i,s:s, s:[I]}",
 		        "id", 41,
 		        "method", "mining.suggest",
-		        "params", ckp->mindiff);
+		        "params", (int64_t)ckp->mindiff);
 	ret = send_json_msg(cs, req);
 	json_decref(req);
 	if (!ret) {
